% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbquartR-package.R
\docType{package}
\name{mbquartR-package}
\alias{mbquartR-package}
\alias{mbquartR}
\title{Finding Manitoba, Canada, quarter sections made easier}
\description{
\code{mbquartR} is an R package for locating quarter sections in Manitoba using lat/long coordinates
or the legal land descriptions using the  Manitoba Original Survey Legal Descriptions data set from the Data MB website (\url{https://geoportal.gov.mb.ca/})
There are four main functions in this package:
}
\details{
\enumerate{
\item Download the Manitoba Original Survey Legal Descriptions data set
\itemize{
\item \code{\link[=quarters_dl]{quarters_dl()}}
}
\item Search for legal land descriptions using lat and long coordinates
}
\itemize{
\item \code{\link[=search_coord]{search_coord()}}
}
\enumerate{
\item Search lat and long coordinates for using legal land descriptions
}
\itemize{
\item \code{\link[=weather_interp]{weather_interp()}}
}
\enumerate{
\item Plot search results on an interactive map
}
\itemize{
\item \code{\link[=map_quarter]{map_quarter()}}
}
}
\author{
\strong{Maintainer}: Alex Koiter \email{koitera@brandonu.ca} (\href{https://orcid.org/0000-0002-9355-9561}{ORCID})

}
