% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mbquartR_example}
\alias{mbquartR_example}
\title{Manitoba Original Survey Legal Descriptions data}
\format{
\subsection{\code{mbquartR_example}}{

A data frame with 4 rows and 14 columns:
\describe{
\item{FID}{Unique ID number}
\item{Formal Legal Description}{The formal legal description
(Leading zeros on numbers)}
\item{Quarter}{Quarter Section value}
\item{SECTION}{Section value}
\item{TOWNSHIP}{Township value}
\item{RANGE}{Range value}
\item{x, y}{Coordinates}
...
}
}
}
\source{
\url{https://geoportal.gov.mb.ca/}
}
\usage{
mbquartR_example
}
\description{
A subset of data from Manitoba Original Survey Legal Descriptions data
}
\keyword{datasets}
