% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_coord.R
\name{search_coord}
\alias{search_coord}
\title{Find coordinates}
\usage{
search_coord(long, lat)
}
\arguments{
\item{long}{Longitude in decimal degrees}

\item{lat}{Latitude in decimal degrees}
}
\value{
A data frame
}
\description{
Find coordinates
}
\examples{
x <- data.frame(long = c(-101.4656, -99.99768), lat = c(51.81913, 49.928926))
search_coord(long = x$long, lat = x$lat)
search_coord(long = c(-101.4656, -99.99768), lat = c(51.81913, 49.928926))
}
