% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{quarters_dl}
\alias{quarters_dl}
\title{Download the MB quarter section data}
\usage{
quarters_dl(force = FALSE, ask = TRUE)
}
\arguments{
\item{force}{Whether to force a download of the data}

\item{ask}{Whether to ask to create the cache folder for MB quarter section data}
}
\value{
A .csv file of Manitoba Original Survey Legal Descriptions and coordinates
}
\description{
Download the MB quarter section data
}
\examples{
\dontrun{
quarters_dl()
}

}
