% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_legal.R
\name{search_legal}
\alias{search_legal}
\title{Find the coordinates of a MB quarter section}
\usage{
search_legal(x)
}
\arguments{
\item{x}{Vector of quarter section legal land descriptions}
}
\value{
A tibble of latitude and longitude coordinates
}
\description{
Find the coordinates of a MB quarter section
}
\examples{
search_legal(x = c("NE-11-33-29W", "SW-20-2-1W"))
}
